+++ src/database/seeds/01-users.js
@@ -1,6 +1,8 @@
 const User = require('../../models/User.js')
+const { Model } = require('objection')
 
-let dataRows=[
+
+let userDataRows = [
   {email: 'usertest@mail.com', password: 'wordpass'},
   {email: 'hello@mail.com', password: 'password'},
   {email: 'travo@mail.com', password: 'wowow'}
@@ -8,10 +10,10 @@ let dataRows=[
 
 exports.seed = function(knex, Promise) {
   // Deletes ALL existing entries
+  Model.knex(knex)
   return knex('user').del()
     .then(function () {
-      console.log(User  )
       // Inserts seed entries
-      return User.query().insert(dataRows);
-    });
+      return Promise.all(userDataRows.map( u => User.query().insert(u) ))
+    })
 };